Initial commit -- cal-3040 from Osmosian.com
This commit is a copy of the official version as it was on
March 5, 2014 (in www.osmosian.com/cal-3040.zip).
It is a complete development environment for Plain English
programming including an iconoclastic interface, a simplified file
manager, an elegant text editor, a hexadecimal dumper, a Win32/x86
native-code-generating compiler/linker, and a wysiwyg page layout
facility for documentation.  A 65,000 word lexicon is provided for
spell-checking; the lexicon is not used by the compiler since the
vocabulary and much of the grammar of a Plain English program are
defined by the programmer.  This commit includes the native version
of the documentation.  (PDFs can be generated directly from within
the integrated development environment.)

This commit does not include the PDF version of the documentation.
Although the documentation provides instructions for writing
the sample program, this commit does not include the sample
program as a distinct file.

Both the PDF documentation and the sample program
are available as part of the official version's zip download.

The PDF version of the documentation is also available at
http://www.osmosian.com/instructions.pdf