Made main window not fill up whole screen; organized working folder.
Made the main window about 1000 pixels wide.
Made the main window about 50 pixels shorter than the screen,
so that the operating system's taskbar can appear at the bottom.
The pixel counts are approximate, because the program assumes
that Windows thinks there are 96 pixels per inch
(regardless of what Windows thinks the pixel density is).
The program still does not have any resize functionality,
and still might not yield control merely because the user
clicks outside of its display area.

Changed the name of the executable (used for editing code)
from cal-3040.exe to cal-3041.exe.
Moved the source code to a draft subfolder,
which is where the draft executable is built.