Added regression testing, and a tool for displaying test results.
This feature stores the results of regression tests
in a structure attached to the status button.
The tests are run as the last step of the program's
initialization.  If all tests pass, the status button
is tinted lightest green.  If any test(s) fail,
the status button is tinted lightest red.
The status button returns to white the next time
a program is compiled or run.

Clicking on the status button re-tints the status button,
and shows more details about the test results.
Specifically, it brings up a menu item summarizing
how many tests passed and/or failed.  Clicking the
menu item brings up a similar summary, followed by
summaries of up to 5 tests.  If all tests passed,
the 5 tests are the first 5 tests that were run.
If any tests failed, the 5 tests are the first
5 failing tests.  The number of summaries can
be adjusted using a global variable.