Fixed memory allocation/deallocation for status' command.
In order to use submenus to show the regression test
results, there needs to be a command.  Because a command
is a thing (an element of a doubly linked list),
the command needs to be manually allocated and destroyed.

Made the command belong to the status,
so it is initialized by the status' initializer,
and destroyed by the status' finalizer.

Updated .gitignore to ignore WINE-generated dump(s),
such as those caused if the status' command is unallocated.

Also highlighted the current font height
in the font height menu.