In regression test results, adjusted font height and refactored.
If the current font height
(which is user chosen on a tab-by-tab basis)
is smaller than the default font height,
use the current font height when displaying
the regression test results.

Reduced duplicate code in the event handlers
that display the regression test results.

Reduced duplicated code when summarizing how many
tests passed and/or failed.