Started writing unit tests for the talker.  Improved the test results' grammar.
Wrote a failing unit test for the talker.  The test tries to get
the class ID for "winmm.dll".  Presumably said class ID can be gotten
on Windows.  Currently said class ID cannot be gotten on a basic WINE
system.  (Wine also has a winmm.dll, but it is not registered
in Wine's registry by default.)  The test fails either way,
because it does not know what to expect.

Dumbed down the test result summaries:
If one test fails, that does not necessarily mean all tests failed.
Therefore started such sentences with the number, not "All" or "The".

Began work to let user copy a test result summary to the clipboard,
by clicking on the test result.  This work is not complete yet,
so such a click will not be handled.

Began work to let the user change the number of test results
that can be shown.  This work is not complete yet,
so such a click will not be handled.