extracted 'the status' from 'the desktop'.
The status feature includes the regression test framework
(but not most of the regression tests themselves),
plus the code to summarize the regression test results
and display sample test results as a menu.
The status feature also includes the feature (in CAL-3040)
that displays the result of the most recent action.

The only regression tests in 'the status'
are to test the features in the file.