extracted 'the colors' from 'the noodle'.
'The colors' includes the colors themselves,
plus the most of the code for manipulating the colors.

'The colors' does not include any brushes or pens,
nor any code for drawing strings, lines or shapes.
There is one exception:  Windows' null pen and null brush are included,
because they are in a procedure that mostly defines colors.