extracted 'characters' from 'the noodle'.
'characters' includes the definitions of the bytes, the keys,
and their respective pointers.  It does not include nibbles or numbers.
The file also includes much of the code that manipulates bytes and keys.
Any manipulations that require Windows API calls are in other file(s).
Most manipulations involving nibbles or strings are in other file(s).