refactored the menu handler switch so yank, font height, and test result do not need every choice hard coded in the switch.
Added code to the desktop, so that the menu handler switch can call shorter
menu handler switch functions for yank, font height, and test result's
handle ... given ... menu handlers.  The shorter switches extract
the relevant number or height from the handlers, and in turn call
the relevant menu handler procedures.

This refactoring allows the number of test results that can be displayed
to only be limited by the screen size, and allows each test result
to be handled differently.  (For example, to be copied to the clipboard.)

This refactoring also allows the choices for font heights and yank distances
to be set dynamically.  (For example, the user can enter an "other" font height,
and CAL will be able to add the new font height to the menu.  Similarly,
the yank distances in the menu can depend on the number of twips per pixel.)