Made regression tests wait until user clicks the status the first time.
It is possible for the regression tests to fail, and never return a result.
(e.g., with a null pointer exception, or with an infinite loop.)
To allow the user to see that the compilation succeeded quickly,
removed the regression tests from the program initialization.

Instead, the first time that the user clicks on the status button,
the regression tests will run.  The regression tests will cache
their results, and show their results as before.